refactor: LottoExecutor를 통해 역할 분리
